requirements.txt
beanie
fastapi
python-dotenv
uvicorn